refactor: change function names and improve folder structure
