+++ src/controllers/auth/signup/signup.controller.ts
@@ -1,4 +1,4 @@
-import { createUserAuthData } from '../../../services/auth/create-user';
+import { signUpFactory } from '../../../services/auth/signup/signup';
 import { SingUpController } from '../../../interfaces/controllers.interfaces';
 import UserAuthRepositoryFactory from '../../../repositories/prisma/prisma-user-auth.repository';
 
@@ -7,7 +7,7 @@ export default (function signUpController(): SingUpController {
     async exec(req, res, next) {
       try {
         const userAuthRepository = UserAuthRepositoryFactory();
-        const signUpService = createUserAuthData(userAuthRepository);
+        const signUpService = signUpFactory(userAuthRepository);
         const newUser = await signUpService.exec(req.body);
         return res.status(201).json({
           user: newUser,
